Tacho: make int-int default

+++ packages/shylu/shylu_node/tacho/src/Tacho.hpp
@@ -43,8 +43,9 @@ typedef int size_type;
 typedef int ordinal_type;
 typedef size_t size_type;
 #else
+// default is int_int
 typedef int ordinal_type;
-typedef size_t size_type;
+typedef int size_type;
 #endif
 
 ///
